requirements.txt
asgiref==3.7.2
attrs==20.3.0
Babel==2.13.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
blinker==1.6.2
Brlapi==0.8.4
cachetools==5.3.2
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
colorama==0.4.6
command-not-found==0.3
cryptography==38.0.4
cupshelpers==1.0
datadog-api-client==2.15.0
dbus-python==1.3.2
defer==1.0.6
distlib==0.3.8
distro==1.8.0
distro-info==1.5
Django==4.2.4
django-cors-headers==4.3.0
django-model-utils==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
dtrx==8.4.0
filelock==3.13.1
flasgger==0.9.7.1
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.8
Flask-Login==0.6.2
google-api-core==2.17.1
google-api-python-client==2.119.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
greenlet==2.0.1
gunicorn==20.1.0
gyp==0.1
html5lib==1.1
httplib2==0.20.4
idna==2.6
importlib-metadata==4.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.1
jeepney==0.8.0
jellyfish==0.11.2
Jinja2==3.1.3
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
keyring==23.9.3
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lib50==3.0.7
louis==3.24.0
lxml==4.9.2
Mako==1.2.4.dev0
markdown-it-py==2.1.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistune==3.0.1
more-itertools==8.10.0
mypy==1.5.1
mypy-extensions==1.0.0
mysql-connector-python==8.0.15
mysqlclient==2.2.0
netifaces==0.11.0
oauthlib==3.2.2
olefile==0.46
onboard==1.4.1
packaging==23.1
parameterized==0.9.0
passlib==1.7.4
pexpect==4.8.0
Pillow==9.4.0
pipenv==2023.11.15
platformdirs==4.1.0
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycodestyle==2.6.0
pycups==2.0.1
pydbus==0.6.0
Pygments==2.14.0
PyGObject==3.44.1
PyJWT==2.6.0
pymongo==4.5.0
pyparsing==3.0.9
python-apt==2.5.3+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
pytz==2023.3.post1
pyxdg==0.28
PyYAML==6.0
redis==5.0.1
referencing==0.30.2
reportlab==3.6.12
requests==2.18.4
requests-oauthlib==1.3.1
rich==13.3.1
rpds-py==0.9.2
rsa==4.9
SecretStorage==3.3.3
six==1.16.0
soupsieve==2.4
SQLAlchemy==1.4.46
sqlparse==0.4.4
submit50==3.1.1
systemd-python==235
termcolor==1.1.0
typing_extensions==4.7.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
unity-tweak-tool==0.0.7
uritemplate==4.1.1
urllib3==1.22
usb-creator==0.3.16
uwidgets==1.0.0
vboxapi==1.0
virtualenv==20.25.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.3.6
xdg==5
xkit==0.0.0
xlrd==1.2.0
zipp==1.0.0
